refactor(rest): merge RoutingTable into HttpHandler
Initially, the RoutingTable was created to leave router complexity out
of the HttpHandler class. This has its own problems too: most
of HttpHandler's public API was just delegating the implementation
to RoutingTable, which was making API changes cumbersome.

After recent changes that extracted route registration and lookup into
standalone RestRouter classes, there is no longer need for this split
and the two classes can be joined together.
