+++ benchmark/package.json
@@ -35,6 +35,7 @@
     "src"
   ],
   "dependencies": {
+    "@loopback/context": "^1.0.0",
     "@loopback/example-todo": "^1.0.2",
     "@loopback/openapi-spec-builder": "^1.0.0",
     "@loopback/rest": "^1.1.0",
